pyproject.toml
[project]
name = "failure-data-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.9.0",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "google-auth>=2.40.2",
    "google-cloud-bigquery>=3.33.0",
    "google-cloud>=0.34.0",
    "google-genai>=1.16.1",
    "gradio>=5.31.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pytest-asyncio>=0.23.8",
    "pytest>=7.4.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.41",
]
